// src/Our.Umbraco.HeadlessPreview/Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ApiResponse = {
    isSuccess: boolean;
};

export type ConfigurationApiResponse = {
    isSuccess: boolean;
    data: (PreviewConfiguration);
};

export type EventMessageTypeModel = 'Default' | 'Info' | 'Error' | 'Success' | 'Warning';

export type NotificationHeaderModel = {
    message: string;
    category: string;
    type: EventMessageTypeModel;
};

export type PreviewConfiguration = {
    templateUrl: string;
    previewModeSettings: Array<(PreviewModeSettingContentType | PreviewModeSettingNodeId)>;
    disabled: boolean;
    configuredFromSettingsFileOrCode: boolean;
};

export type PreviewMode = 'UseHeadlessPreview' | 'UseStandardPreview' | 'DisablePreview';

export type PreviewModeApiResponse = {
    isSuccess: boolean;
    data: (PreviewModeResponse);
};

export type PreviewModeResponse = {
    previewMode: PreviewMode;
};

export type PreviewModeSettingContentType = {
    type: PreviewModeSettingType;
    contentTypes: Array<(string)>;
    mode: PreviewMode;
};

export type PreviewModeSettingNodeId = {
    type: PreviewModeSettingType;
    nodeIds: Array<(number)>;
    includeDescendants: boolean;
    mode: PreviewMode;
};

export type PreviewModeSettingType = 'ContentType' | 'NodeId';

export type GetConfigurationResponse = ((ConfigurationApiResponse));

export type GetConfigurationError = (unknown);

export type GetPreviewModeData = {
    query?: {
        contentTypeGuid?: string;
        nodeGuid?: string;
    };
};

export type GetPreviewModeResponse = ((PreviewModeApiResponse));

export type GetPreviewModeError = (unknown);

export type SaveConfigurationData = {
    body?: (PreviewConfiguration);
};

export type SaveConfigurationResponse = ((ApiResponse));

export type SaveConfigurationError = (unknown);